Preparing AlrDetector for move in later CL.
This was part of a larger CL to move ownership of alr detector. Since that
cl had issues with lock usage I'll defer that to a CL where TaskQueues are
introduced.

This cl does two things:
1: It reduces the probe controller dependency. This change adds one extra
lock reaquire in the probe controller.

2: It moves interval length calculation inside the alr-detector. This make
the interface simpler for calling code. This means it will not be able to
detect alr until called twice. This increases the detection time with one
period if the application starts in alr-mode.

Bug: webrtc:8415
Change-Id: I7061308d68bb7615f2c085df7c9f03d52fefd8e9