Update patch set 7
Patch Set 7:

(2 comments)

We can't wait for a month.  Is the new format ready to use?  In other words, can we use the new format now and get the info into scientist? If not, I suggest we work with the old format and then when all the other things are moved over to the new format, the ICE events also be moved over.

Patch-set: 7
Reviewer: Gerrit User 5660 <5660@58829da1-049c-39fb-b951-ebdcd0984767>