Update patch set 9
Patch Set 9:

(11 comments)

In addition to consolidating the naming convention, the major update gives each candidate pair a numeric id and organizes networking related info to the description struct, which is only logged when a candidate pair is added for the first time. The change to the analyzer code so far reuses the existing charting utilities, but I found a compatibility issue in the frontend when a large number of candidate pairs are generated by some calls, the chart cannot be rendered because of failure to generating the color palette. Since we will likely add new types of charts for networking related logs, and changes in the analyzer feel like a placeholder. How about dropping the changes related to charting in this CL, and following up with CLs extending both the frontend and charting in the analyzer?

Patch-set: 9
Reviewer: Gerrit User 5887 <5887@58829da1-049c-39fb-b951-ebdcd0984767>