Update patch set 15
Patch Set 15:

(7 comments)

In addition to edits addressing the comments, I have written the interface of IceEventLog and removed the logging method for the config events. It seems there is possibility in the future that the description for candidate pairs should be included in some non-config events, so I left the handling of config and non-config (in the sense of ICE logging not RtcEventLog) in the implementation and let the logger implementation, along with the proto schema, decides if the description message should be included. The caller of the ice logger always includes the description of a candidate pair and doesn't have to distinguish between config and non-config events. This change impacts the other CL.

Patch-set: 15
Reviewer: Gerrit User 5053 <5053@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5887 <5887@58829da1-049c-39fb-b951-ebdcd0984767>