Update patch set 16
Patch Set 16:

I have added a method to dump metadata/description of candidate pairs in batch to the log buffer to mitigate one remaining concern that logged config events carrying metadata may not get serialized due to contention with other RtcEvents, buffer capacity and output file size limit. This method is called in the other CL when ICE is complete. This change basically adds possibly duplicate info to the log to increase the likelihood of metadata being serialized to the file, instead of applying the bigger change that we include the metadata in all events. We may find better solutions in following CLs.

Patch-set: 16
Reviewer: Gerrit User 5887 <5887@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0