Update patch set 2
Patch Set 2:

(1 comment)

It's unclear to me what's the purpose of this change. TaskQueue is not an interface, so you don't gain much flexibility by injection. Will the TaskQueue somehow be shared with other code outside of RtcEventLog? Why?

The TaskQueue is injected with ownership transfer, which makes things simpler (if tasks accessing the RtcEventLog are posted to the TaskQueue, then the TaskQueue must be destroyed before anything accessed by the tasks).

Patch-set: 2
CC: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>