Update patch set 1
Patch Set 1:

I had a look at the approach mentioned with grpc. I noticed that this is implemented in the X509_check_host function in 1.0.2 so I have decided to go for that. Let me know if there is anything else you want addressed for this bug.

Patch-set: 1
Reviewer: Gerrit User 6930 <6930@58829da1-049c-39fb-b951-ebdcd0984767>
Work-in-progress: false