Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> We should add unit test for this validation

It appears in the past this entire path has zero coverage. I am going to try find a way to correctly test it but it is a bit difficult being a private member of an implementation of an interface.

Patch-set: 1
Reviewer: Gerrit User 6930 <6930@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0