Improve text logs in the network stack.
1) The local network info is added to the candidate pair string
presentation in logs so that we do not need the crosss reference to the
logs of candidate gathering to find out the network where the local
candidate resides. 2) All connectivity check transmission
logs are promoted to the INFO level. In the past, logs of connectivity
checks for a candidate pair is demoted to the VERBOSE level once it has
strong connectivity. This may hide potentially abnormal behavior before
the candidate pair degrades to a weaker connectivity, which takes a
non-trivial period by default.

Bug: None
Change-Id: Iaa337394cad803515e26e254814aa04ed2213eab