Update patch set 5
Patch Set 5:

(4 comments)

Thanks a lot for the work! I have some inline comments plus a general thought: can we generalize the concept (not necessarily the logic) here a bit.

We may think it this way: weather to decode existing packets after expansion or not depends on whether we would expect more packets to arrive within target delay. So that is to let packet_buffer_ to provide the end timestamp. Yes, PacketBuffer has a method NumSamplesInBuffer(), but as you noticed that it does not count DTX well, both in the case that the last packet is DTX, and the case that a middle packet is DTX.

Can we try to make NumSamplesInBuffer() method better and let DecisionLogic take, in stead of, dtx_future (bool), something more generic (e.g., end timestamp of PacketBuffer) as input (we may let DecisionLogic take the whole PacketBuffer as input).

Patch-set: 5
Reviewer: Gerrit User 7015 <7015@58829da1-049c-39fb-b951-ebdcd0984767>