Update patch set 12
Patch Set 12:

> Patch Set 12:
> 
> Hello, everyone,
> 
> I've noticed this CL only now and it seems like it might have the same issue as boost asio library (https://github.com/chriskohlhoff/asio/issues/287). In addition to check android api version it looks like it necessary to check if HAVE_PTHREAD_COND_TIMEDWAIT_MONOTONIC defined. I'm not really sure if the problem is there, but decided to leave a comment to let you know there was a problem with pthread_cond_timedwait_relative_np.
> 
> Thank you for attention,
> Yura.

I wrote this cl under the assumption that 

1. all android versions we care about have either pthread_condattr_setclock or pthread_cond_timedwait_monotonic_np, and that

2. the former function (which I'd prefer to use) is available if and only if __ANDROID_API__ >= 21. 

It's unfortunate if we ever have to fallback to gettimeofday. (Which we'd need to do in case __ANDROID_API < 21 and pthread_cond_timedwait_monotonic_np isn't accessible in the sdk). But should be reasonably easy to do, if needed.

We'll see if any of our test bots fails compilation.

Patch-set: 12