Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Hello, Sami,
> > 
> > Could you please review this CL which adds additional handling of --verbose flag passed to build_aar.py. If it is necessary to open an issue on bugs.webrtc.org just tell me, I'll create it.
> > 
> > Thanks in advance,
> > Yura.
> 
> Hi,
> 
> Could you clarify what is the use case for this? To me it seems that if there are problems with these commands, it would be easy try to run the command manually with the flag enabled?

I use this script on build server and for some reason ninja compilation fails on it, but not locally. I could probably do some work on build server to replace build_aar.py with manual gn/ninja invocation, but it could diverge with implementation of build_aar.py which I prefer to use. I've noticed that there is a --verbose switch in build_aar.py but it is not passed into programs it called (gn, ninja), so I decided to propose fixing this, since it might be useful not only for me.

Patch-set: 1
Reviewer: Gerrit User 6305 <6305@58829da1-049c-39fb-b951-ebdcd0984767>