Update patch set 1
Patch Set 1:

My main worry that this might be too verbose for a lot of debugging scenarios.

Perhaps this could be implemented instead in a way where arbitrary parameters to gn or ninja can be passed instead. This would solve your problem and give a lot of flexibility for usage.

Patch-set: 1
Reviewer: Gerrit User 5112 <5112@58829da1-049c-39fb-b951-ebdcd0984767>