Additional switches to build_aar.py
Extra switches to GN could be passed via --extra-gn-switches.
Extra switches to Ninja could be passed via --extra-ninja-switches.
They could be used in different scenarios, when additional switches
need to be passed to GN or Ninja. For example, when diagnosing
build issues extra switch `-v` could be passed to enable
verbose logging of GN and Ninja.

Bug: None
Change-Id: I09d18a57b3df4e698784fb7d58c02e8adecddefa