Update patch set 28
Patch Set 28:

> Patch Set 28:
> 
> Nice!  Are the below numbers correct? Seems like much more than just 25K?
> 
> 4.095 MiB	3.966 MiB	−131.69 KiB	−3.1406%	apks/AppRTCMobile.apk	android_arm64_rel
> 3.345 MiB	3.184 MiB	−164.20 KiB	−4.7943%	libjingle_peerconnection_datachannelonly_so.so	android_arm64_rel
> 7.065 MiB	6.685 MiB	−388.43 KiB	−5.3694%	libjingle_peerconnection_so.so	android_arm64_rel
> 3.705 MiB	3.608 MiB	−99.43 KiB	−2.6207%	apks/AppRTCMobile.apk	android_rel
> 2033.82 KiB	1943.73 KiB	−90.09 KiB	−4.4297%	libjingle_peerconnection_datachannelonly_so.so	android_rel
> 4.439 MiB	4.242 MiB	−201.16 KiB	−4.4259%	libjingle_peerconnection_so.so	android_rel
> 20.023 MiB	19.913 MiB	−112.57 KiB	−0.5490%	WebRTC.framework/WebRTC	ios_api_framework
> 45.746 MiB	43.624 MiB	−2.122 MiB	−4.6377%	obj/libwebrtc.a	linux_rel

Jonas observes -25 kB locally for libjingle_peerconnection_so.so on arm32 opt. From a discussion with phoglund@, it seems like the larger numbers here are caused by the bots building with DCHECKs enabled. That doesn't cause a large error in the size *diff* normally, but this CL changes how [D]CHECKs are implemented, so in addition to getting credited for shrinking the ~1000 CHECK calls, it also gets credit for shrinking the ~6000 DCHECK calls.

Patch-set: 28