[desktopCapture Mac]reorder execution order in start/release processing
This cl is to move the RegisterRefreshAndMoveHandlers to be done on
capture thread, and reverse some execution order of releasing processing,
also remove a lock since the handler is on capturing thread too.
As we doubt the existing sequence may be the cause of a crash due to
race conditions at end of capture.

Bug:851883
Change-Id: I2254a69815144415424a77b4c82f150cfc369585