Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'm not familiar enough with the sdp standard so will let others answer that.
> A unit test covering at least the problem we found would be nice though :)

I agree that unit tests would be desirable. All these functions are static at the moment, I guess to not expose to the public interface of webrtcsdp. Do you have any best practices for situations like these. I know one design that I've seen elsewhere is to declare an XXXXinternal.h or XXXXhelper.h and put the declarations of the functions there and then move the definitions to a helper namespace. By doing that the functions doesn't clutter the interface of the class but are still possible to test.

Patch-set: 1
Reviewer: Gerrit User 7641 <7641@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0