Update patch set 1
Patch Set 1:

Agree that this needs an unit test.

Suggest that you mangle one of the simpler tests like IceCredentialsInCandidateStringIgnored (I picked on that one because it is a successful parse) and add attributes to taste:

a=name\0
a=name\r\n
a=name:and then some stuff
a=namethatshouldnotmatch

The function you're testing is about 4 call levels down in the stack, but that's how the APIs are constructed.

Note: From RFC 4566, it's not clear that a space can ever happen - the next element is either a value with : in front of it or the end of the line.

Patch-set: 1
Reviewer: Gerrit User 5634 <5634@58829da1-049c-39fb-b951-ebdcd0984767>