Deliver partial SCTP messages when the SID is (unexpectedly) changed.
We're receiving an SCTP messages one chunk at a time. The sender is supposed to set the MSG_EOR bit on the last chunk of a message. A crash (RTC_CHECK) happened when the sender starts a new message without indicate the end of the previous message. This is likely due to an SCTP implementation that doesn't (correctly) support the MSG_EOR bit.

This change, rather than calling RTC_CHECK, delivers partial messages when SID is (unexpectedly) changed.

Bug: chromium:884926
Change-Id: I18c773bb60ae724b735a83933eedd030f6360a3c