Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

Code change lgtm. I'm not that familiar with the call tests, though. Are you confident that 10ms isn't so short that it will make tests flaky? 

But it should be easy to increase it a factor 10 or so if it turns out to cause problems.

Patch-set: 1
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=+1