Allow use of CroppingWindowCapturer via CreateWindowCapturer
Currently, apps using WebRTC for window capture only get the benefits of using
CroppingWindowCapturer on Windows (described below) after changing calls to
DesktopCapturer::CreateWindowCapturer to instead call
CroppingWindowCapturer::CreateCapturer. This change removes that requirement,
making the faster capture-screen-and-crop path available via the older & more
discoverable API.

This was proposed previously on https://crbug.com/webrtc/10734#c2 , and in code
review discussion around the fix for that bug it was suggested that this change
be made separately.

Benefits of using CroppingWindowCapturer's capture-screen-and-crop path when
possible:
1) It's significantly faster, up to ~36ms/frame (~160x) faster than the capture-
window-contents path in my testing (more details are in the bug). This
difference increased with the recent fix for https://crbug.com/webrtc/10734 .
2) It allows capture of menus & tooltips (plus dialogs if
https://crbug.com/webrtc/10767 is fixed), partially mitigating
https://crbug.com/980864 .

An example of an app/browser that could benefit from this change is Firefox,
which currently calls DesktopCapturer::CreateWindowCapturer.

This change also makes CroppingWindowCapturer::CreateCapturer respect the
detect_updated_region flag if set in the options it's passed on Windows.

Bug: webrtc:10825
Change-Id: Ib77e5facc7240c5df311fe1fe204d0d8ea22a96a