Replace DCHECKS with logging to avoid incorrect metrics activation
This CL replaces the current DCHECKS for avoiding multiple
enabling of metrics to instead use logging.
The reasons for that are that
1) The DCHECK will only catch this in debug builds
2) Currently, the DCHECK causes the code after the DCHECK
   to not be tested in debug builds.
3) The impact of activating the metrics twice is actually none.

Beyond this change, the CL also changes the use of the flag
g_rtc_histogram_called that checks if histograms were reported
before the metrics were enabled to use a logging instead of a
DCHECK.

That reasons for that using logging tests in DEBUG mode will
be able to test the code after the DCHECK (which is running in
release builds and therefore should be tested).

Bug: None
Change-Id: I6ebfc098bdf74de2b32b001fc22530f5a41956f3