Enable use rtc::SystemTimeNanos() provided by Chromium
This is the third CL out of three to enable overriding
of the function SystemTimeNanos() in rtc_base/system_time.cc

When WebRTC is built as part of Chromium the rtc::SystemTimeNanos()
function provided by Chromium will be used. This is controlled
by the build argument rtc_exclude_system_time which directly
maps to the macro WEBRTC_EXCLUDE_SYSTEM_TIME.

By doing this we are making sure that the WebRTC and Chromium
clocks are the same.

Bug: chromium:516700
Change-Id: If7f749c4aadefb1cfc07ba4c7e3f45dc6c31118b