Check for BMI2 support before enabling AVX2 code paths
On Windows, compiling with AVX2 support will also result in generation
of BMI2 instructions. Some early Haswell CPUs reportedly support AVX2
but not BMI2. We have seen crashes (illegal instruction) on these CPUs
in our AVX2 code paths on MULX instructions (part of BMI2).

Including a check for BMI2 when checking for AVX2 support is
expected to solve the issue.

Bug: chromium:1315519
Change-Id: I3a0a9838f1f632704ba505ecbb81a6f8b1889319